Move morale and related functions to Character
