+++ src/savegame_json.cpp
@@ -510,6 +510,8 @@ void Character::load( JsonObject &data )
         }
     }
 
+    morale->load( data );
+
     _skills->clear();
     JsonObject pmap = data.get_object( "skills" );
     for( const std::string &member : pmap.get_member_names() ) {
@@ -544,6 +546,16 @@ void Character::load( JsonObject &data )
     // Bionic power should not be negative!
     if( power_level < 0_mJ ) {
         power_level = 0_mJ;
+    }
+
+    JsonArray overmap_time_array = data.get_array( "overmap_time" );
+    overmap_time.clear();
+    while( overmap_time_array.has_more() ) {
+        point pt;
+        overmap_time_array.read_next( pt );
+        time_duration tdr = 0_turns;
+        overmap_time_array.read_next( tdr );
+        overmap_time[pt] = tdr;
     }
 }
 
@@ -612,6 +624,9 @@ void Character::store( JsonOut &json ) const
     if( is_mounted() ) {
         json.member( "mounted_creature", g->critter_tracker->temporary_id( *mounted_creature ) );
     }
+
+    morale->store( json );
+
     // skills
     json.member( "skills" );
     json.start_object();
@@ -629,6 +644,16 @@ void Character::store( JsonOut &json ) const
         json.member( "power_level", units::to_kilojoule( power_level ) );
     }
     json.member( "max_power_level", units::to_kilojoule( max_power_level ) );
+
+    if( !overmap_time.empty() ) {
+        json.member( "overmap_time" );
+        json.start_array();
+        for( const std::pair<point, time_duration> &pr : overmap_time ) {
+            json.write( pr.first );
+            json.write( pr.second );
+        }
+        json.end_array();
+    }
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -724,16 +749,6 @@ void player::store( JsonOut &json ) const
         json.end_object();
     }
     json.end_array();
-
-    if( !overmap_time.empty() ) {
-        json.member( "overmap_time" );
-        json.start_array();
-        for( const std::pair<point, time_duration> &pr : overmap_time ) {
-            json.write( pr.first );
-            json.write( pr.second );
-        }
-        json.end_array();
-    }
 }
 
 /**
@@ -844,15 +859,6 @@ void player::load( JsonObject &data )
         bcdata.read( "pos", bcpt );
         camps.insert( bcpt );
     }
-    JsonArray overmap_time_array = data.get_array( "overmap_time" );
-    overmap_time.clear();
-    while( overmap_time_array.has_more() ) {
-        point pt;
-        overmap_time_array.read_next( pt );
-        time_duration tdr = 0_turns;
-        overmap_time_array.read_next( tdr );
-        overmap_time[pt] = tdr;
-    }
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -915,8 +921,6 @@ void avatar::store( JsonOut &json ) const
     json.member( "guts", guts );
 
     json.member( "translocators", translocators );
-
-    morale->store( json );
 
     // mission stuff
     json.member( "active_mission", active_mission == nullptr ? -1 : active_mission->get_id() );
@@ -1038,8 +1042,6 @@ void avatar::load( JsonObject &data )
     data.read( "guts", guts );
 
     data.read( "translocators", translocators );
-
-    morale->load( data );
 
     std::vector<int> tmpmissions;
     if( data.read( "active_missions", tmpmissions ) ) {
